javadoc and some saftey options for ResourceCopier

+++ metron-platform/metron-test-utilities/src/main/java/org/apache/metron/test/utils/ResourceCopier.java
@@ -26,35 +26,71 @@ import java.nio.file.Path;
 import java.nio.file.SimpleFileVisitor;
 import java.nio.file.attribute.BasicFileAttributes;
 
+/**
+ * Utility class to copy a whole directory.  This is useful for copying files from
+ * src/test/resources to target/testbed for example
+ */
 public class ResourceCopier {
 
-  public static void copyResources(Path sourcePath, Path targetPath) throws IOException{
-    Files.walkFileTree(sourcePath, new SimpleFileVisitor<Path>() {
+  /**
+   * Copy the resources from sourcePath to targetPath. If a resource exists, it will be overwritten.
+   * This is the equivolent of calling <code>copyResources(source,target,true)</code>
+   *
+   * @param sourcePath source {@link Path}
+   * @param targetPath target {@link Path}
+   */
+  public static void copyResources(final Path sourcePath, final Path targetPath)
+      throws IOException {
+    synchronized (ResourceCopier.class) {
+      copyResources(sourcePath, targetPath, true);
+    }
+  }
 
-      @Override
-      public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs)
-          throws IOException {
+  /**
+   * Copy the resources from sourcePath to targetPath. The overwrite flag determines if existing
+   * resources are overwritten or not.
+   *
+   * @param sourcePath source {@link Path}
+   * @param targetPath target {@link Path}
+   * @param overwrite if true, the the target will be overwritten with {@link
+   * java.nio.file.StandardCopyOption#REPLACE_EXISTING}
+   */
+  public static void copyResources(final Path sourcePath, final Path targetPath,
+      final boolean overwrite) throws IOException {
+    synchronized (ResourceCopier.class) {
+      Files.walkFileTree(sourcePath, new SimpleFileVisitor<Path>() {
 
-        Path relativeSource = sourcePath.relativize(dir);
-        Path target = targetPath.resolve(relativeSource);
+        @Override
+        public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs)
+            throws IOException {
 
-        Files.createDirectories(target);
+          Path relativeSource = sourcePath.relativize(dir);
+          Path target = targetPath.resolve(relativeSource);
 
-        return FileVisitResult.CONTINUE;
+          Files.createDirectories(target);
 
-      }
+          return FileVisitResult.CONTINUE;
 
-      @Override
-      public FileVisitResult visitFile(Path file, BasicFileAttributes attrs)
-          throws IOException {
+        }
 
-        Path relativeSource = sourcePath.relativize(file);
-        Path target = targetPath.resolve(relativeSource);
+        @Override
+        public FileVisitResult visitFile(Path file, BasicFileAttributes attrs)
+            throws IOException {
 
-        Files.copy(file, target, REPLACE_EXISTING);
+          Path relativeSource = sourcePath.relativize(file);
+          Path target = targetPath.resolve(relativeSource);
 
-        return FileVisitResult.CONTINUE;
-      }
-    });
+          if (overwrite) {
+            Files.copy(file, target, REPLACE_EXISTING);
+          } else {
+            if (!target.toFile().exists()) {
+              Files.copy(file, target);
+            }
+          }
+
+          return FileVisitResult.CONTINUE;
+        }
+      });
+    }
   }
 }
